* move CommandDefinition in a new package (model) * the shellbuilder adds now the command-definitions to the shell
